Enable reconnecting after a network disruption
For example, when the wifi connection is lost and regained the WebSocket would
disconnect and not get reconnected.

Adds a reconnect() function to Pusher.

Additionally, the Channel needs to unsubscribe regardless of the connected
state to make sure the Pusher delists the Channel and that reauthentication
happens when reconnecting.

Bump to version 1.1.0 (due to new API).